rename: user → member 패키지명 변경
