+++ src/main/java/team05/integrated_feed_backend/module/member/controller/MemberControllerDocs.java
@@ -1,11 +1,11 @@
-package team05.integrated_feed_backend.module.user.controller;
+package team05.integrated_feed_backend.module.member.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import team05.integrated_feed_backend.common.BaseApiResponse;
-import team05.integrated_feed_backend.module.user.dto.request.MemberSignupReq;
+import team05.integrated_feed_backend.module.member.dto.request.MemberSignupReq;
 
 @Tag(name = "Member", description = "멤버 관련 API")
 public interface MemberControllerDocs {
