+++ src/main/java/team05/integrated_feed_backend/module/member/entity/enums/MemberStatus.java
@@ -1,0 +1,5 @@
+package team05.integrated_feed_backend.module.member.entity.enums;
+
+public enum MemberStatus {
+	UNVERIFIED, VERIFIED
+}
