+++ src/main/java/team05/integrated_feed_backend/module/member/event/SignedUpEvent.java
@@ -1,0 +1,8 @@
+package team05.integrated_feed_backend.module.member.event;
+
+import team05.integrated_feed_backend.module.member.entity.Member;
+
+public record SignedUpEvent(
+	Member member
+) {
+}
