+++ src/main/java/team05/integrated_feed_backend/module/member/event/publisher/MemberEventPublisher.java
@@ -1,4 +1,4 @@
-package team05.integrated_feed_backend.module.user.event.publisher;
+package team05.integrated_feed_backend.module.member.event.publisher;
 
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.scheduling.annotation.Async;
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import team05.integrated_feed_backend.module.user.entity.Member;
-import team05.integrated_feed_backend.module.user.event.SignedUpEvent;
+import team05.integrated_feed_backend.module.member.entity.Member;
+import team05.integrated_feed_backend.module.member.event.SignedUpEvent;
 
 @Component
 @RequiredArgsConstructor
