+++ src/main/java/team05/integrated_feed_backend/module/member/repository/MemberRepository.java
@@ -1,10 +1,10 @@
-package team05.integrated_feed_backend.module.user.repository;
+package team05.integrated_feed_backend.module.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import team05.integrated_feed_backend.module.user.entity.Member;
-import team05.integrated_feed_backend.module.user.entity.vo.Email;
+import team05.integrated_feed_backend.module.member.entity.Member;
+import team05.integrated_feed_backend.module.member.entity.vo.Email;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
