--- src/main/java/team05/integrated_feed_backend/module/user/entity/enums/MemberStatus.java
@@ -1,5 +1,0 @@
-package team05.integrated_feed_backend.module.user.entity.enums;
-
-public enum MemberStatus {
-	UNVERIFIED, VERIFIED
-}
