--- src/main/java/team05/integrated_feed_backend/module/user/event/SignedUpEvent.java
@@ -1,8 +1,0 @@
-package team05.integrated_feed_backend.module.user.event;
-
-import team05.integrated_feed_backend.module.user.entity.Member;
-
-public record SignedUpEvent(
-	Member member
-) {
-}
