+++ src/test/java/team05/integrated_feed_backend/module/member/entity/vo/EmailTest.java
@@ -1,4 +1,4 @@
-package team05.integrated_feed_backend.module.user.entity.vo;
+package team05.integrated_feed_backend.module.member.entity.vo;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.ThrowableAssert.*;
